+++ CHANGELOG.md
@@ -16,6 +16,7 @@ details.
 - Support to process rules from multiple locations in`checkrules.js`.
 - Optional/Custom function to process rules data retrieved from the remote
   source.
+- Support rule prefixes.
 
 ### Changed
 - Use `chalk` for colorful display of the result.