+++ CHANGELOG.md
@@ -11,7 +11,7 @@ details.
 
 ---
 
-## [Unreleased]
+## [1.3.0] - 2017-01-20
 ### Added
 - Support to process rules from multiple locations in`checkrules.js`.
 - Optional/Custom function to process rules data retrieved from the remote